Crate nss

Modules

aes

cert

ec

ecdh

pbkdf2

pk11

pkixc

secport

Macros

scoped_ptr

Structs

Error

Enums

ErrorKind

Functions

ensure_initialized

ensure_initialized_with_profile_dir

Use this function to initialize NSS if you want to manage keys with NSS. ensure_initialized_with_profile_dir initializes NSS with a profile directory (where key4.db will be stored) and appropriate flags to persist keys (and certificates) in its internal PKCS11 software implementation. If it has been called previously with a different path, it will fail. If ensure_initialized has been called before, it will also fail.

Type Aliases

Result