nss::pk11
Module sym_key

Functions

authenticate_with_primary_password

Authorize NSS key store against a user-provided primary password. Only available with the keydb feature.

authentication_with_primary_password_is_needed

Check weather a primary password has been set and NSS needs to be authenticated. Only available with the keydb feature.

get_or_create_aes256_key

Retrieve a key, identified by name, from the internal NSS key store. If none exists, create one, persist, and return. Only available with the keydb feature.

hkdf_expand